requirements.txt
sqladmin
python-dotenv
python-jose
annotated-types==0.7.0
anyio==4.11.0
exceptiongroup==1.3.0
fastapi==0.118.0
idna==3.10
pydantic==2.11.10
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
